Release candidate v3

+++ src/main/java/group5/hotelms/dao/HotelDAO.java
@@ -3,12 +3,16 @@ package group5.hotelms.dao;
 import group5.hotelms.model.Hotel;
 import java.util.Set;
 
+/**
+ * @author Andey Ponomarenko
+ */
 public interface HotelDAO {
     boolean saveHotel(Hotel hotel);
 
     boolean deleteHotel(Hotel hotel);
 
-    Hotel getHotelById(long id);
+    Hotel getHotelById(int id);
 
     Set<Hotel> getAllHotels();
+
 }
